% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_pop.r
\name{add_esp}
\alias{add_esp}
\title{add_esp}
\usage{
add_esp(df, age_band_col)
}
\arguments{
\item{df}{Dataframe containing a column with 5 year age bands.}

\item{age_band_col}{A column found in df containing age bands.}
}
\value{
(dataframe) A dataframe with a new column containing the european standard
population values for each age band.
}
\description{
This function creates a new column of the European standard population
in 5 year age bands to match 5 year age bands in your dataframe. Your dataframe should
already have a column called 'age_band' with 5 year age bands starting from 0 and the
age bands should be labelled in the following format: 0, 5, 10, 15, 20, 25, 30, 35, 40, 45,
50, 55, 60, 65, 70, 75, 80, 85, 90.
}
\examples{
test_df <- data.frame(id = 1:6, age_band = c(0, 5, 10, 15, 35, 50))

df <- add_esp(df = test_df,
              age_band_col = age_band)

}
