% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_band_for_phe.r
\name{age_band_for_phe}
\alias{age_band_for_phe}
\title{age_band_for_phe}
\usage{
age_band_for_phe(df, age_column)
}
\arguments{
\item{age_column}{A column found in 'data' containing ages.}

\item{data}{Dataframe containing a column with single year ages.}
}
\value{
(dataframe) A dataframe with a new column containing the age bands.
}
\description{
This function creates a new column of 5 year age bands based on a
column of single year ages. The age bands are labelled with the lower bound of
each band to match the PHEindicators package input requirements.
}
\examples{
test_df <- data.frame(id = 1:6, age = c(0, 3, 4, 5, 12, 91))

df <- age_band_for_phe(data = test_df,
                            age_column = age)


}
