% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10_code_to_name.r
\name{icd10_code_to_name}
\alias{icd10_code_to_name}
\title{icd10_code_to_name}
\usage{
icd10_code_to_name(df, code_column, detailed = TRUE)
}
\arguments{
\item{code_column}{A column found in 'data' containing ICD 10 codes.}

\item{detailed}{Specifies whether to return detailed or broad categories. Defaults to TRUE for detailed categories. Set to FALSE for broad categories.}

\item{data}{Dataframe containing a column with ICD 10 codes.}
}
\value{
(dataframe) A dataframe with a new column containing the ICD 10 code names.
}
\description{
This function creates a new column of death categories based on a column of ICD 10 codes
}
\examples{
df_death <- icd10_code_to_name(data = death_data,
                            code_column = underlying_cause_of_death,
                            detailed = TRUE)

to get broad categories 
df_death <- icd10_code_to_name(data = death_data,
                            code_column = underlying_cause_of_death,
                            detailed = FALSE)

}
