% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpho_colours.r
\name{kpho_colours}
\alias{kpho_colours}
\title{kpho_colours}
\usage{
kpho_colours(geography)
}
\arguments{
\item{geography}{String of the type of geography to get the colour scheme for. Options include: \code{"district"}, \code{"hcp"}, \code{"nhs_trust"}, \code{"pcn"}, \code{"utla_national"} and \code{"district_medway"}}
}
\value{
A patchwork object.
}
\description{
This function prints standard colour schemes for different geography types plotted on KPHO charts.
}
\examples{
kpho_colours("district")

}
