% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpho_trend.r
\name{kpho_trend}
\alias{kpho_trend}
\title{kpho_trend}
\usage{
kpho_trend(
  data,
  values,
  timeperiods,
  group,
  description,
  stat_type,
  timerange,
  palette = NULL,
  ci = FALSE
)
}
\arguments{
\item{data}{Dataframe containing a long-form dataset of a health indicator.}

\item{values}{A unique variable found in 'data' containing numeric values to be plotted.}

\item{timeperiods}{A unique variable found in 'data' which list the different timeperiods which the values are calculated for.}

\item{group}{A unique variable found in 'data' to group the value outputs on the graph.}

\item{description}{A string describing what the values show.}

\item{stat_type}{A string describing the units of the values presented.}

\item{timerange}{A concatenated vector of years or timeperiods to plot on the trend graph.}

\item{palette}{A concatenated string of group names and color values to be applied to the graph. Default is set to null. Use \code{kpho_colours()}.}

\item{ci}{A logical variable for the inclusion of confidence intervals which defaults to FALSE. If set to TRUE, the columns 'lowercl' and 'uppercl' in 'data' are needed.}
}
\value{
A patchwork object.
}
\description{
This function creates a trend graph of a specific health indicator using ggplot and patchwork.
}
\examples{
smoking_trend <- kpho_trend(data = smoking_prevalence,
                            values = "numbers",
                            timeperiods = "years",
                            group = "area",
                            description = "Smoking prevalence by district in Kent"
                            stat_type = "Prevalence (\%)"
                            timerange = c("2012", "2013", "2014", "2015"),
                            ci = FALSE)

}
