% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable.r
\name{lt}
\alias{lt}
\title{lt}
\usage{
lt(df)
}
\arguments{
\item{df}{A dataframe with two columns: single year of age and the associated deaths within a certain time period. Note: all ages, even if there are 0 deaths recorded, must be present for an accurate calculation of life expectancy.}
}
\value{
A dataframe with all constituent elements of a lifetable calculation.
}
\description{
This function creates a lifetable of a given dataframe of deaths by single year of age. It does not accurately calculate life expectancy for those below age 1. Please see the ONS guidance for how to caluclate quarterly life expectancy for those below age 1.
}
